requirements.txt
selenium
python-dotenv
beautifulsoup4
requests 